Rename c_ota.h to _subroutines/c_ota.h
